use long for error code

+++ src/main/java/com/eveonline/api/exceptions/ApiResultException.java
@@ -22,16 +22,16 @@ package com.eveonline.api.exceptions;
  */
 public class ApiResultException extends ApiException {
 
-	private int errorCode;
+	private long errorCode;
 	private String errorMessage;
 
-	public ApiResultException(int errorCode, String errorMessage) {
+	public ApiResultException(long errorCode, String errorMessage) {
 		super(errorCode + ": " + errorMessage);
 		this.errorCode = errorCode;
 		this.errorMessage = errorMessage;
 	}
 
-	public int getErrorCode() {
+	public long getErrorCode() {
 		return errorCode;
 	}
